// src/interfaces/TiposProducto/TiposProducto.ts
// Generated by https://quicktype.io

export interface TiposProductoResponse {
    status:  number;
    message: string;
    data:    DataTipoProducto[];
}

export interface DataTipoProducto {
    id:              number;
    tipoProducto:    string;
    descripcion:     string;
    usuarioCreacion: string;
    fechaCreacion:   string;
    estadosId:       number;
}